FPGA/driver_controller.sv: blanking, gclk and sclk
The module driver_controller now output the gclk and sclk signal.
Because we must wait for the end of a gclk segment (which is 2^N wide,
in our case 512 bit wide), we have an internal blanking while we must
driev sclk to 0 to avoid losing data. Indeed sclk is used to shif the
common shift register, thus while we wait for the end of gclk segment we
don't send data into it and don't want to shift it.