FPGA/src/framebuffer.sv: pad with 0 after 16 bits
Add an rgb_counter to know which color we are sending, and wires to
compute the correct bit address in the voxel. Since we use 6 bits for
the green it is not a simple modulo. Instead we use a COLOR_BASE
localparam with the strating bits of the three colors (0,5,11) and add
to it a color_idx that goes from 5 to 0. When the color is not green we
substract 1 to this color_idx so it goes from 4 to 0.