FPGA/src/framebuffer.sv: change frame buffer size
The framebuffer were to big to fit in the FPGA.
Now the framebuffer only holds a column for each driver, reducing his
size by 8 times. Hence we have to swap buffer each time we change the
multiplexing, and change a little bit the way we read the ram and read
the buffer.