FPGA/tb_src/systemc/driver_controller: init testbench with reset Improve monitor so as to run some processes when special events occur. For example, start a new process to check configuration protocol each time a FCWRTEN occurs on the LAT. Some things are still missing in the Stimuler part, including signals between Simuler and Monitor so as to check data.
Add protocol check. It forces the driver controller to finish the full
drawing as soon as it started to draw. It is not asked by the driver but
gives a good and simple protocol check.

Use of sc_join: as it is running in a thread with a sensibility
list, it should use wait_clocked() instead of wait().