FPGA/src: add new clock to match timing
The drivers have non trivial timing for the lat and data, thus we need a
high frequency clock to respect them. I have added a new clock (which
will probably be 66 MHz) and derive sclk and gclk from it. For the
counters (hence the driver state) the clock is in phase quadrature with
sclk thus we send data/lat within requirement.