FPGA/src/framebuffer.sv: change comments, buffer structure and addresses
As explained in the wiki and the framebuffer comments, the panel are
facing each other, thus they display the same thing at the same time, so
we only need 40 columns in a slice. When the panels will have made half
a turn, they will display the 40 other column, so we still have 80
pixels wide resolution. However this implies that the framebuffer remembers
that we have made half a turn to send the correct data, or that we change
the generation of the slices.
In order for the framebuffer to remain position agnostic, the later was
chosen. The SBC needs to send even columns for 128 slices, and then odd
columns for 128 other. It also needs to reverse the odd columns to take
into account the rotation of the panel.