FPGA/src/rgb_logic.sv: add synchronization signals
Add stream_ready signal, which is outpu when enough slices have been
written in ram to begin the display.
Add rgb_enable input, which is input by the SPI module when we should
start writing in ram. When rgb_enable drives high, the rgb_logic module
waits for vsync to be low before writing, as we want to start at the
beginning of a frame.