FPGA/framebuffer.sv: LUT endianness, writing index and swap buffer issue
The last voxel in the buffer wasn't write because the has_reached_end check
was done one cycle too soon.

The LUTs declaration (color, driver, base address) written with little
endianess in mind but stored in big endian as any normal systemverilog variable.
I changed the declaration to store in little endian to keep the LUT easy to
read.

The buffer were swapped too soon, a new signal column_sent help to know
precisely when they need to be swap.