FPGA/src/framebuffer.sv: read next slice when waiting for it
Fill the read buffer with the first column in the next slice when
waiting for it, so it is ready to display when position_sync comes.
This cannot be done for the very first slice, meaning that the very
first column will not be displayed, creating a little artifact at the
beggining of the stream. However at 30 fps we should no see it.