FPGA/tb_src/systemc/tb_rgb_logic: fix tests for rgb logic
Fix tests for rgb_logic, adding the following checks:
+ memory addresses are correctly incremented or reset in the normal
  behaviour
+ data written in the RAM is correct
+ rgb_enable and write_enable signals have the correct behaviour
+ in case of incomplete frame, memory addr is correctly reset
Now, testbench generate data at negedge(clk), because it is expected to
be treated at posedge by the rgb_logic, so that it can generate its own
data at negedge.
This commit also fix checkRAM next address check, which was not using the fact that
ramAddr can cycle to 0 after a while and needed complex modulo and
subtractions to work properly with the previous model. Now the code is
simpler and less error-prone.