PFGA/src/rgb_logic.sv: fix sampling, signals and frame reset
This commit adds the following:
+ fix ram not being reset at incomplete frame
+ sample addr at negedge and fix end_of_RAM signal
+ fix is_end_of_RAM using the wrong signal

ram_addr will be sampled at posedge in the ram module, thus it needs
to be modified at negedge in the rgb_logic module.
Moreover, there was a bug in the is_end_of_RAM signal. It stemmed from
the fact it wasn't written like combinatorial logic with an assign.
Finally, is_end_of_RAM must use the ram_addr counter instead of the
pixel_counter, because it is checking a RAM property.