FPGA/src/spi.sv: use comb logic to process shift reg on last cycle
A bug in the SPI module design led to having the last byte of the config
processed later, because it needed one cycle more of spi_clk to do it.
By moving to combinatorial logic for the shift register comparison, we
can avoid the need of this additional cycle and do the decoding at the
exact time the new byte is entirely received.